requirements.txt
aiogram~=3.22.0
SQLAlchemy[asyncio]~=2.0.45
aiosqlite~=0.20.0
python-dotenv~=1.2.1
tzdata~=2024.1
